+++ src/views/InputComp.vue
