+++ src/views/Styleguide.vue
@@ -3,15 +3,15 @@
     <div>{{ $t('styleguide') }}!</div>
     <p class="font-bold text-red-300">styleguide</p>
 
-    <input-test />
+    <input-comp />
   </div>
 </template>
 
 <script>
-import InputTest from './InputTest.vue';
+import InputComp from '@/views/InputComp.vue';
 
 export default {
-  components: { InputTest },
+  components: { InputComp },
   name: 'Styleguide',
 };
 </script>
